refactor: auditlog fe (#213)
* refactor: auditlog fe

* build: 1.1.1

* remove: license file
