+++ backend/src/zelthy/apps/shared/tenancy/templates/app_panel.html
@@ -17,7 +17,7 @@
   <script>
     var csrf_token = "{{ csrf_token }}";
   </script>
-  <script src="{% static 'app_panel/js/build.v1.1.0.min.js' %}"></script>
+  <script src="{% static 'app_panel/js/build.v1.1.1.min.js' %}"></script>
   <script src="{% static 'js/jquery/3.7.1/jquery.min.js' %}"></script>
   {% include 'session_security/all.html' %}
 </body>
