+++ frontend/src/metadata.json
@@ -1,1 +1,1 @@
-{ "buildMajor": 1, "buildMinor": 1, "buildPatch": 0, "buildTag": "" }
+{"buildMajor":1,"buildMinor":1,"buildPatch":1,"buildTag":""}
