--- frontend/src/pages/appAccessLogs/components/Table/RowMenu.jsx
@@ -1,42 +1,0 @@
-import { Menu, Transition } from '@headlessui/react';
-import { Fragment, useState } from 'react';
-import { usePopper } from 'react-popper';
-import { ReactComponent as TableRowKebabIcon } from '../../../../assets/images/svg/table-row-kebab-icon.svg';
-
-export default function RowMenu({ rowData }) {
-	const [referenceElement, setReferenceElement] = useState(null);
-	const [popperElement, setPopperElement] = useState(null);
-	const { styles, attributes } = usePopper(referenceElement, popperElement, {
-		placement: 'bottom-end',
-		modifiers: [
-			{
-				name: 'offset',
-				options: {
-					offset: [16, 8],
-				},
-			},
-		],
-	});
-
-	return (
-		<Menu as="div" className="relative flex">
-			<Menu.Button
-				className="flex w-full justify-center focus:outline-none"
-				ref={(ref) => setReferenceElement(ref)}
-			>
-				<TableRowKebabIcon className="text-[#5048ED]" />
-			</Menu.Button>
-			<Transition
-				as={Fragment}
-				// @ts-ignore
-				ref={(ref) => setPopperElement(ref)}
-				style={styles['popper']}
-				{...attributes['popper']}
-			>
-				<Menu.Items className="absolute top-[30px] right-0 w-[186px] origin-top-right rounded-[4px] bg-white shadow-table-menu focus:outline-none">
-					<div className="p-[4px]"></div>
-				</Menu.Items>
-			</Transition>
-		</Menu>
-	);
-}
