+++ frontend/src/pages/appApplicationObjectsLogs/components/AppApplicationObjectsLogs/index.js
@@ -1,22 +1,15 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import { ReactComponent as AddUserIcon } from '../../../../assets/images/svg/add-user-icon.svg';
 import useApi from '../../../../hooks/useApi';
 import BreadCrumbs from '../../../app/components/BreadCrumbs';
 import {
-	openIsAddNewUserModalOpen,
 	selectAppApplicationObjectsLogsData,
 	selectAppApplicationObjectsLogsTableData,
+	selectIsAppApplicationObjectsLogsDataEmpty,
 	selectRerenderPage,
 	setAppApplicationObjectsLogsData,
-	selectIsAppApplicationObjectsLogsDataEmpty,
 } from '../../slice';
-import ActivateUserModal from '../Models/ActivateUserModal';
-import AddNewUserModal from '../Models/AddNewUserModal';
-import DeactivateUserModal from '../Models/DeactivateUserModal';
-import EditUserDetailsModal from '../Models/EditUserDetailsModal';
-import ResetPasswordModal from '../Models/ResetPasswordModal';
 import Table from '../Table';
 
 export default function AppApplicationObjectsLogs() {
@@ -33,10 +26,6 @@ export default function AppApplicationObjectsLogs() {
 	);
 
 	const dispatch = useDispatch();
-
-	const handleAddNewUser = () => {
-		dispatch(openIsAddNewUserModalOpen());
-	};
 
 	function updateAppApplicationObjectsLogsData(value) {
 		dispatch(setAppApplicationObjectsLogsData(value));
@@ -87,18 +76,6 @@ export default function AppApplicationObjectsLogs() {
 			<div className="flex grow flex-col gap-[20px]">
 				<div className="flex items-center justify-between py-[12px] pl-[40px] pr-[48px]">
 					<BreadCrumbs />
-					{/* {isAppApplicationObjectsLogsDataEmpty ? null : (
-						<button
-							type="button"
-							onClick={handleAddNewUser}
-							className="flex gap-[8px] rounded-[4px] bg-primary px-[16px] py-[7px]"
-						>
-							<span className="font-lato text-[14px] font-bold leading-[20px] text-[#FFFFFF]">
-								New User
-							</span>
-							<AddUserIcon />
-						</button>
-					)} */}
 				</div>
 				<div className="flex grow flex-col overflow-x-auto">
 					{isAppApplicationObjectsLogsDataEmpty ? (
@@ -107,31 +84,13 @@ export default function AppApplicationObjectsLogs() {
 								<h3 className="first-app-text font-source-sans-pro text-[64px] font-[700] leading-[72px]">
 									application objects logs
 								</h3>
-								{/* <p className="font-source-sans-pro text-[18px] font-semibold leading-[24px] text-[#212429]">
-									description to be added
-								</p> */}
 							</div>
-							{/* <button
-								type="button"
-								onClick={handleAddNewUser}
-								className="flex gap-[8px] rounded-[4px] bg-primary px-[16px] py-[7px]"
-							>
-								<span className="font-lato text-[14px] font-bold leading-[20px] text-[#FFFFFF]">
-									Create New User
-								</span>
-								<AddUserIcon />
-							</button> */}
 						</div>
 					) : appApplicationObjectsLogsData ? (
 						<Table tableData={appApplicationObjectsLogsData?.audit_logs} />
 					) : null}
 				</div>
 			</div>
-			<AddNewUserModal />
-			<EditUserDetailsModal />
-			<DeactivateUserModal />
-			<ActivateUserModal />
-			<ResetPasswordModal />
 		</>
 	);
 }
