--- frontend/src/pages/appAuditLogs/routes/index.js
@@ -1,11 +1,0 @@
-import { Navigate, Route, Routes } from 'react-router-dom';
-import AppAuditLogs from '../components/AppAuditLogs/index.js';
-
-export const AppAuditLogsRoutes = () => {
-	return (
-		<Routes>
-			<Route path="" element={<AppAuditLogs />} />
-			<Route path="*" element={<Navigate to="." />} />
-		</Routes>
-	);
-};
