+++ frontend/src/pages/appFrameworkObjectsLogs/components/AppFrameworkObjectsLogs/index.js
@@ -1,22 +1,16 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import { ReactComponent as AddUserIcon } from '../../../../assets/images/svg/add-user-icon.svg';
 import useApi from '../../../../hooks/useApi';
 import BreadCrumbs from '../../../app/components/BreadCrumbs';
 import {
 	openIsAddNewUserModalOpen,
 	selectAppFrameworkObjectsLogsData,
 	selectAppFrameworkObjectsLogsTableData,
+	selectIsAppFrameworkObjectsLogsDataEmpty,
 	selectRerenderPage,
 	setAppFrameworkObjectsLogsData,
-	selectIsAppFrameworkObjectsLogsDataEmpty,
 } from '../../slice';
-import ActivateUserModal from '../Models/ActivateUserModal';
-import AddNewUserModal from '../Models/AddNewUserModal';
-import DeactivateUserModal from '../Models/DeactivateUserModal';
-import EditUserDetailsModal from '../Models/EditUserDetailsModal';
-import ResetPasswordModal from '../Models/ResetPasswordModal';
 import Table from '../Table';
 
 export default function AppFrameworkObjectsLogs() {
@@ -87,18 +81,6 @@ export default function AppFrameworkObjectsLogs() {
 			<div className="flex grow flex-col gap-[20px]">
 				<div className="flex items-center justify-between py-[12px] pl-[40px] pr-[48px]">
 					<BreadCrumbs />
-					{/* {isAppFrameworkObjectsLogsDataEmpty ? null : (
-						<button
-							type="button"
-							onClick={handleAddNewUser}
-							className="flex gap-[8px] rounded-[4px] bg-primary px-[16px] py-[7px]"
-						>
-							<span className="font-lato text-[14px] font-bold leading-[20px] text-[#FFFFFF]">
-								New User
-							</span>
-							<AddUserIcon />
-						</button>
-					)} */}
 				</div>
 				<div className="flex grow flex-col overflow-x-auto">
 					{isAppFrameworkObjectsLogsDataEmpty ? (
@@ -107,31 +89,13 @@ export default function AppFrameworkObjectsLogs() {
 								<h3 className="first-app-text font-source-sans-pro text-[64px] font-[700] leading-[72px]">
 									framework objects logs
 								</h3>
-								{/* <p className="font-source-sans-pro text-[18px] font-semibold leading-[24px] text-[#212429]">
-									description to be added
-								</p> */}
 							</div>
-							{/* <button
-								type="button"
-								onClick={handleAddNewUser}
-								className="flex gap-[8px] rounded-[4px] bg-primary px-[16px] py-[7px]"
-							>
-								<span className="font-lato text-[14px] font-bold leading-[20px] text-[#FFFFFF]">
-									Create New User
-								</span>
-								<AddUserIcon />
-							</button> */}
 						</div>
 					) : appFrameworkObjectsLogsData ? (
 						<Table tableData={appFrameworkObjectsLogsData?.audit_logs} />
 					) : null}
 				</div>
 			</div>
-			<AddNewUserModal />
-			<EditUserDetailsModal />
-			<DeactivateUserModal />
-			<ActivateUserModal />
-			<ResetPasswordModal />
 		</>
 	);
 }
