+++ frontend/src/routes/index.js
@@ -1,12 +1,12 @@
+import { useEffect } from 'react';
+import { Toaster } from 'react-hot-toast';
+import { useDispatch } from 'react-redux';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
+import useApi from '../hooks/useApi';
 import { PlatformAppRoutes } from '../pages/app/routes';
 import { PlatformRoutes } from '../pages/platform/routes';
-import { PlatformUserManagementRoutes } from '../pages/platformUserManagement/routes';
-import toast, { Toaster } from 'react-hot-toast';
-import { useEffect } from 'react';
-import useApi from '../hooks/useApi';
 import { setAppPanelInitialData } from '../pages/platform/slice';
-import { useDispatch } from 'react-redux';
+import { PlatformUserManagementRoutes } from '../pages/platformUserManagement/routes';
 
 export const AppRoutes = () => {
 	const location = useLocation();
