+++ frontend/src/store/index.js
@@ -1,20 +1,17 @@
 import { configureStore } from '@reduxjs/toolkit';
-
-import platformReducer from '../pages/platform/slice';
-import platformUserManagementReducer from '../pages/platformUserManagement/slice';
-import appUserRolesReducer from '../pages/appUserRoles/slice';
-import appUserManagementReducer from '../pages/appUserManagement/slice';
-import appTaskManagementReducer from '../pages/appTaskManagement/slice';
+import appChatbotReducer from '../pages/app/slice';
+import appApplicationObjectsLogsReducer from '../pages/appApplicationObjectsLogs/slice';
+import appConfigurationReducer from '../pages/appConfiguration/slice';
+import appFrameworkObjectsLogsReducer from '../pages/appFrameworkObjectsLogs/slice';
+import appPackagesManagementReducer from '../pages/appPackagesManagement/slice';
 import appPermissionsManagementReducer from '../pages/appPermissionsManagement/slice';
 import appPoliciesManagementReducer from '../pages/appPoliciesManagement/slice';
-import appConfigurationReducer from '../pages/appConfiguration/slice';
+import appTaskManagementReducer from '../pages/appTaskManagement/slice';
 import appThemeConfigurationReducer from '../pages/appThemeConfiguration/slice';
-import appPackagesManagementReducer from '../pages/appPackagesManagement/slice';
-import appChatbotReducer from '../pages/app/slice';
-import appAuditLogsReducer from '../pages/appAuditLogs/slice';
-import appApplicationObjectsLogsReducer from '../pages/appApplicationObjectsLogs/slice';
-import appFrameworkObjectsLogsReducer from '../pages/appFrameworkObjectsLogs/slice';
-import appAccessLogsReducer from '../pages/appAccessLogs/slice';
+import appUserManagementReducer from '../pages/appUserManagement/slice';
+import appUserRolesReducer from '../pages/appUserRoles/slice';
+import platformReducer from '../pages/platform/slice';
+import platformUserManagementReducer from '../pages/platformUserManagement/slice';
 
 export default configureStore({
 	reducer: {
@@ -29,9 +26,7 @@ export default configureStore({
 		appThemeConfiguration: appThemeConfigurationReducer,
 		appPackagesManagement: appPackagesManagementReducer,
 		appChatbot: appChatbotReducer,
-		appAuditLogs: appAuditLogsReducer,
 		appApplicationObjectsLogs: appApplicationObjectsLogsReducer,
 		appFrameworkObjectsLogs: appFrameworkObjectsLogsReducer,
-		appAccessLogs: appAccessLogsReducer,
 	},
 });
